rearranging to run offline for D2

+++ make_plots.py
@@ -333,28 +333,39 @@ def run():
         # Calculate xrange and yrange for histograms
         Xrange = [grid['xpsi'].min(), grid['xpsi'].max()]
         Yrange = [grid['ypsi'].min(), grid['ypsi'].max()]
+
+        # Histogram of starting locations
+        if whichtype == 'cross': # results are in xg, yg
+            xp0, yp0, _ = tracpy.tools.interpolate2d(d['xg0'], d['yg0'], grid, 'm_ij2xy')
+        elif 'coast' in whichtype:  # results are in xp, yp
+            xp0 = d['xp0']; yp0 = d['yp0']
+
     elif whichtype == 'D2': # results are in xg, yg
         d = netCDF.Dataset(Files[0][0])
 
         # Calculate xrange and yrange for histograms in ll
         Xrange = [grid['lonpsi'].min(), grid['lonpsi'].max()]
         Yrange = [grid['latpsi'].min(), grid['latpsi'].max()]
-    # Histogram of starting locations
-    if whichtype == 'cross': # results are in xg, yg
-        xp0, yp0, _ = tracpy.tools.interpolate2d(d['xg0'], d['yg0'], grid, 'm_ij2xy')
-    elif 'coast' in whichtype:  # results are in xp, yp
-        xp0 = d['xp0']; yp0 = d['yp0']
-    elif whichtype == 'D2': # results are in xg, yg
+    
+        # Histogram of starting locations
         # xp0, yp0 are lonp, latp in this case
         xp0, yp0, _ = tracpy.tools.interpolate2d(d.variables['xg'][:,0], d.variables['yg'][:,0], grid, 'm_ij2ll')
-    # pdb.set_trace()
-    # For D2 and fsle, Hstart contains indices of drifters seeded in bins
-    Hstart, xe, ye = calc_histogram(xp0, yp0, whichtype, bins=bins, Xrange=Xrange, Yrange=Yrange)
-    if whichtype == 'D2':
-        xe, ye = grid['basemap'](xe, ye) # change from lon/lat
+
+        # Histogram of starting locations
         Hstartfile = 'calcs/dispersion/hist/Hstart_bins' + str(bins[0]) + '.npz'
-        np.savez(Hstartfile, Hstart=Hstart, xe=xe, ye=ye) 
-        
+
+    if os.path.exists(Hstartfile): # just read in info
+        Hstartf = np.load(Hstartfile)
+        xe = Hstartf['xe']; ye = Hstartf['ye']
+        Hstart = Hstartf['Hstart']
+    else:
+        # For D2 and fsle, Hstart contains indices of drifters seeded in bins
+        Hstart, xe, ye = calc_histogram(xp0, yp0, whichtype, bins=bins, Xrange=Xrange, Yrange=Yrange)
+
+        if whichtype == 'D2':
+            xe, ye = grid['basemap'](xe, ye) # change from lon/lat
+            np.savez(Hstartfile, Hstart=Hstart, xe=xe, ye=ye) 
+
 
     d.close()
